+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Added support for preserving existing throughput buckets during SQL container throughput updates.
 
 ## Version 2.0.7
 * Added support for throughput buckets.