+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Fixed issue where the restore command failed to detect the source account, even when it existed.
 
 ## Version 2.0.7
 * Added support for throughput buckets.