+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -25,6 +25,9 @@
 ## Version 2.0.7
 * Added support for throughput buckets.
 
+## Version 1.18.2
+* Fixed issue where the restore command failed to detect the source account, even when it existed.
+
 ## Version 1.18.1
 * Fixed incorrect help documents for MongoDB cmdlets.